evaluate only a single pair of numbers at a time
Example: you enter a number (12), followed by an operator button (+),
a second number button (7), and a second operator button (-).
The calculator should then do the following: first, evaluate the initial
pair of numbers (12 + 7), then display the result of that calculation (19).
Finally, use that result (19) as the first number in a new calculation,
along with the next operator (-)